Moved MUI background image to its own component, started working on theme provider changes
