+++ client/src/components/App.jsx
@@ -12,27 +12,9 @@ import Login from './Login'; // import Login component
 import InputForm from './InputForm';
 import SignUp from './SignUp';
 import About from './About';
+import Background from './Background';
 
-import backgroundImage from '../assets/images/background1.jpg';
-
-const BackgroundBox = styled(Box)(({ theme }) => ({
-  position: 'relative',
-  minHeight: '100vh',
-  width: '100%',
-  '&::before': {
-    content: '""',
-    position: 'absolute',
-    top: 0,
-    left: 0,
-    right: 0,
-    bottom: 0,
-    backgroundImage: `url(${backgroundImage})`,
-    backgroundSize: 'cover',
-    backgroundPosition: 'center',
-    opacity: 0.5, // Adjust this value to change transparency (0 is fully transparent, 1 is fully opaque)
-    zIndex: -1,
-  },
-}));
+import theme from '../themes/customTheme';
 
 const ContentContainer = styled(Box)({
   position: 'relative',
@@ -46,7 +28,7 @@ console.log('App.jsx is running');
 function App() {
   return (
     <div>
-      <BackgroundBox>
+
         <ContentContainer>
           <Router>
             <Nav />
@@ -59,7 +41,7 @@ function App() {
             </Routes>
           </Router>
         </ContentContainer>
-      </BackgroundBox>
+
     </div>
   );
 }
